Removed scripts that are no longer required, moved json_fields_to_csv to wrangler repo and updated relevant readmes
